Change GITObjectHash so that it passes the tests again.
It seems that since adding BridgeSupport to the framework the Foundation framework is included by MacRuby when the tests are run, this seemed to break *a lot* of GITObjectHash. These have now been fixed by changing the PACK Data encoding to NSISOLatin1StringEncoding. I'm not sure why this fixes it, but it could just be that Array#pack returns 'H*' packed strings in ISO Latin-1 encoding...

The GITObjectHash -unpackedStringFromString: method has been reverted back so that it does not use the -unpackedDataFromData: method as this was returning erroneous data, again a likely encoding issue. Restoring the StringFromString method to how it was implemented previously has fixed the errors in the tests.