Add GITTree class.
GITTree uses the Parsing category on GITObject to extract the item data from its contents. The GITObject(Parsing) category was originally written by Brian Chapados as part of the CocoaGit project. It utilises a structure definition of the match data to be found in a byte array. It is designed to be more efficient for the purposes of git object string extraction than NSScanner which spends a lot of its time converting items to NSString instances while parsing. The GITObject(Parsing) category was added to CocoaGit in commit 5c0ba8c0a748ad0d4255156e9674297a4dccf156.