Removing .unpackedStringFromString and .packedStringFromString from GITObjectHash
The concept of a "packed" NSString doesn't actually make sense. A packed Hash is 20 byte blob of data. There are no string encodings that represent this. Since NSString is very encoding aware, there is no way to represent a packed hash as an NSString. NSData should be used for this.