BUG FIX: GITTree parsing item mode values.
GITTree when parsing the value of the mode of an item was interpreting the contents as a plain integer rather than the hexadecimal contents which it actually is. This commit changes how this field is processed from the string form into the NSUInteger form it should be. An NSScanner is employed to convert the strings hexadecimal form into its decimal form.