Fixes a reported issue regarding repos created by older hg2git utilities.
Older hg2git utilities don't seem to create the .git/branches directory. This folder does not appear to be used at present, branches are stored in .git/refs/heads. This commit removes .git/branches from the list of skeleton directories which are used in the repo root sanity check. The .git/branches directory is now listed in a separate legacy skeleton directories array which is only used when creating a new repository root.